requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.20
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
importlib-metadata==8.5.0
Markdown==3.7
sqlparse==0.5.3
typing-extensions==4.12.2
zipp==3.20.2
